move type to types.tsx
Типы используемые в приложении следует перенести в файл types.ts

move types to types fix

move types to types 3

types more
