+++ src/components/BurgerIngredients/BurgerIngredientsItem/index.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import Ingredient from "../../../utils/data"
+import Ingredient from "../../../utils/types"
 import { Price } from "../../price"
 import { Counter } from '@ya.praktikum/react-developer-burger-ui-components'
